# docs/SERVER_MODE/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MrRSS is a modern, cross-platform desktop RSS reader with auto-translation, smart feed discovery, and AI-powered summarization.",
        "title": "MrRSS API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/DevXDojo/MrRSS",
            "email": "mail@ch3nyang.top"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.3.27"
    },
    "host": "localhost:1234",
    "basePath": "/api",
    "paths": {
        "/ai/profiles": {
            "get": {
                "description": "Get all AI configuration profiles (without API keys for security)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai-profiles"
                ],
                "summary": "List AI profiles",
                "responses": {
                    "200": {
                        "description": "List of AI profiles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AIProfile"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new AI configuration profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai-profiles"
                ],
                "summary": "Create AI profile",
                "parameters": [
                    {
                        "description": "Profile data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created profile",
                        "schema": {
                            "$ref": "#/definitions/models.AIProfile"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ai/profiles/test-all": {
            "post": {
                "description": "Test all AI profiles' connections and configurations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai-profiles"
                ],
                "summary": "Test all AI profiles",
                "responses": {
                    "200": {
                        "description": "Test results for all profiles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ProfileTestResult"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ai/profiles/test-config": {
            "post": {
                "description": "Test an AI configuration without creating/saving a profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai-profiles"
                ],
                "summary": "Test AI configuration without saving",
                "parameters": [
                    {
                        "description": "Configuration to test",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ProfileTestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test result",
                        "schema": {
                            "$ref": "#/definitions/handlers.ProfileTestResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/ai/search": {
            "post": {
                "description": "Use AI to expand keywords and search articles with relevance ranking",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "AI-powered article search",
                "parameters": [
                    {
                        "description": "Search query",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AISearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "schema": {
                            "$ref": "#/definitions/handlers.AISearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ai/test": {
            "post": {
                "description": "Test AI service configuration (endpoint, API key, model availability)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ai"
                ],
                "summary": "Test AI configuration",
                "responses": {
                    "200": {
                        "description": "Test result (config_valid, connection_success, model_available, response_time_ms, test_time, error_message)",
                        "schema": {
                            "$ref": "#/definitions/handlers.TestResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.AISearchRequest": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
        "handlers.AISearchResponse": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                },
                "error": {
                    "type": "string"
                },
                "search_terms": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "handlers.ProfileRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "custom_headers": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handlers.ProfileTestRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "custom_headers": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                }
            }
        },
        "handlers.ProfileTestResult": {
            "type": "object",
            "properties": {
                "config_valid": {
                    "type": "boolean"
                },
                "connection_success": {
                    "type": "boolean"
                },
                "error_message": {
                    "type": "string"
                },
                "model_available": {
                    "type": "boolean"
                },
                "profile_id": {
                    "type": "integer"
                },
                "profile_name": {
                    "type": "string"
                },
                "response_time_ms": {
                    "type": "integer"
                }
            }
        },
        "handlers.TestResult": {
            "type": "object",
            "properties": {
                "config_valid": {
                    "type": "boolean"
                },
                "connection_success": {
                    "type": "boolean"
                },
                "error_message": {
                    "type": "string"
                },
                "model_available": {
                    "type": "boolean"
                },
                "response_time_ms": {
                    "type": "integer"
                },
                "test_time": {
                    "type": "string"
                }
            }
        },
        "models.AIProfile": {
            "type": "object",
            "properties": {
                "api_key": {
                    "description": "Hidden in responses, only sent when needed",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "custom_headers": {
                    "description": "JSON string of key-value pairs",
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "description": "Default profile for new features",
                    "type": "boolean"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}